Update syntax highlighting with new highlight.min.js
Downloaded a new version of highlight.js 7.3 with
clojure(script) and coffeescript support from:

http://softwaremaniacs.org/soft/highlight/en/download/

For the record, the previous minified highlight.min.js
was found to support html/xml, javascript, and css. To
this list I added clojure and coffeescript. I then
pasted in the new hljs line in by hand so as to preserve
the Showdown stowaway.